Changes that I made to this file that I forgot to commit on my branch.
I suppose my intention was to focus the attention on the fundamental
portion of my original patch but I am still a bit confused as to how I
ended up here.  Anyway it does not really matter since my suggested
patch did end up in the master, so I am just committing this to my
branch for posterity.