Checkpoint scatter work.
All tests pass now that I have determined a way to generalize the test
input parameters to fit the max bailouts and chunk size.

For some reason, switching to the minItemsTestingThreshold length
(from the previous hard-coded value of 13) also fixes the scatter-11
test, but it is not clear to me why this is so.  At this point it is
more important to me to checkpoint my work and move on than to dissect
that oddity.