Put the big array of inputs aside; restrict attention to a smaller set.
The motivation here is largely to make the output smaller and thus
easier to explain/interpret.

Also revised the output formatting so that it will tend to fit in 80
column wide terminal.  (Did this by taking advantage of new fact that
the input array is smaller and thus the reported index takes fewer
columns to report.)