Crazy hacks to get more context during Parallel Bailout failures.
Braindump follows:

As a reminder, here is the usual way Felix has been invoking the js shell:

  TERM=xterm-color PAFLAGS=full IONFLAGS=bailouts ./js

and here is the test case Felix has been using most often:

  (new Matrix([100], ["int32"], function (i) { return i+1; }, {spew:1})).shape

though he has at times varied the size of the shape {10,100,1000}, and
of course toggled the `{spew:1}` argument on and off.

Pushing to github to share with Shu to debug why we are seeing:

  warning: Bailed out of parallel operation: unsupported ArgumentChecks_0 at self-hosted:4104

although we have seen various other scary things such as:

  [Parallel:M]   COMPILE 10316cf38:self-hosted:4072
  [Parallel:M]   END can't compile

potentially in tandem with the addition of a (this === undefined)
condition to a predicate somewhere.